+++ echeck.h
@@ -2,6 +2,8 @@
 #ifndef _ECHECK_H_
 #define _ECHECK_H_
 
+#include <stdlib.h>		/* for size_t */
+
 unsigned GLOBAL_EXIT_ON_FAIL = 0;
 
 int check_char_m(char actual, char expected, const char *msg);
@@ -20,10 +22,10 @@ int check_unsigned_int_m(unsigned int actual, unsigned int expected,
 			 const char *msg);
 int check_unsigned_int(unsigned int actual, unsigned int expected);
 
-int check_byte_array_m(unsigned char *actual, unsigned int actual_len,
-		       unsigned char *expected, unsigned int expected_len,
+int check_byte_array_m(unsigned char *actual, size_t actual_len,
+		       unsigned char *expected, size_t expected_len,
 		       const char *msg);
-int check_byte_array(unsigned char *actual, unsigned int actual_len,
-		     unsigned char *expected, unsigned int expected_len);
+int check_byte_array(unsigned char *actual, size_t actual_len,
+		     unsigned char *expected, size_t expected_len);
 
 #endif /* _ECHECK_H_ */
